+++ begin-react/src/UserList.js
@@ -8,29 +8,12 @@ function User({user}) {
     );
 }
 
-function UserList() {
-    const users = [
-        {
-            id: 1,
-            username: 'abc',
-            email: 'public.abc@gmail.com',
-        },
-        {
-            id: 2,
-            username: 'def',
-            email: 'public.def@gmail.com',
-        },
-        {
-            id: 3,
-            username: 'ghi',
-            email: 'public.ghi@gmail.com',
-        },
-    ];
+function UserList({users}) {
 
     return (
         <div>
             {
-                users.map(user => (<User user={user} key={user.id} />)) // key할 값이 없다면 index 삽입
+                users.map(user => (<User user={user} key={user.id} />)) // key할 값이 없다면 index 삽
             }
         </div>
     );
